Develop

+++ ha-batch/docker/data_purge.sh
@@ -9,6 +9,7 @@
 # 
 # input: MySQL.USER
 # ----------------------------------------------------------------------------------------
+set -eu
 
 echo "------------------------------------------------------------------------"
 echo "START $0"
@@ -16,6 +17,19 @@ echo "------------------------------------------------------------------------"
 
 # call common.sh
 . ./common.sh
+
+# ロック処理
+SCRIPT_NAME=$(basename "$0")
+LOCK_FILE_NAME="${SCRIPT_NAME}.lock"
+LOCK_FILE="${LOCK_DIR}/${LOCK_FILE_NAME}"
+mkdir -p "$(dirname "$LOCK_FILE")"
+
+exec 9>"${LOCK_FILE}"
+
+if ! flock -n 9; then
+  echo "[ERROR] dailyApiLogJob is already running."
+  exit 50
+fi
 
 cd ${BASE_DIR} && docker compose -f docker-compose.yml -f docker-compose.${ENV}.yml run --rm ha-batch --spring.batch.job.name=dataPurgeBatchJob
